Conflicts:
	src/uk/ac/cam/cl/juliet/common/Container.java

+++ src/uk/ac/cam/cl/juliet/common/XDPRequest.java
@@ -1,17 +1,14 @@
 package uk.ac.cam.cl.juliet.common;
 
-public class XDPRequest extends XDPPacket
-{
-	private int[] packetData;
-	
-	public XDPRequest(long id, int[] packetData)
-	{
+public class XDPRequest extends XDPPacket {
+	private byte[] packetData;
+
+	public XDPRequest(long id, byte[] packetData) {
 		super(id);
 		this.packetData = packetData;
 	}
-	
-	public int[] getPacketData()
-	{
+
+	public byte[] getPacketData() {
 		return this.packetData;
 	}
 }
